Create GARBAGE MONITORING SYSTEM
It's a project regarding garbage monitoring system. It consists of three levels of monitoring .If the garbage is at low level it indicates the blue light. If the garbage is at mid level green light glows and if the  trash can is completely filled  red led glows . A message is gone to the municipality by wifi module and they will come and collect the wastes.